Carry sound font ids in 16 bits
Each streamed track brings its own sound font, so a pack with more than
~217 of them pushes font ids past 255. Stored in 8 bits they wrapped onto
vanilla fonts or onto 0xFF, which reads as "no font": the track played
vanilla instruments or nothing at all.

Fixes #7226

+++ soh/soh/resource/type/AudioSequence.h
@@ -12,7 +12,7 @@ typedef struct {
     uint8_t medium;
     uint8_t cachePolicy;
     uint32_t numFonts;
-    uint8_t fonts[16];
+    uint16_t fonts[16];
 } Sequence;
 
 class AudioSequence : public Ship::Resource<Sequence> {
